bot/requirements.txt
aiogram>=3.23.0
python-dotenv>=1.2.1
requests>=2.32.0
qrcode[pil]>=7.4.2
